requirements.txt
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.23
numpy==1.19.5
protobuf==3.14.0
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
